delete lessons and abonements

+++ src/main/java/com/app/service/AbonementServiceImpl.java
@@ -1,9 +1,14 @@
 package com.app.service;
 
+import com.app.dto.response.ApiResponse;
 import com.app.exeption.ResourceNotFoundException;
 import com.app.model.Abonement;
+import com.app.model.ConfirmedLesson;
+import com.app.model.DeletedLesson;
 import com.app.model.Student;
 import com.app.repository.AbonementRepository;
+import com.app.repository.ConfirmedLessonRepository;
+import com.app.repository.DeletedLessonRepository;
 import com.app.repository.StudentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -17,6 +22,8 @@ import java.util.List;
 public class AbonementServiceImpl implements AbonementService{
   private final AbonementRepository abonementRepository;
   private final StudentRepository studentRepository;
+  private final ConfirmedLessonRepository confirmedLessonRepository;
+  private final DeletedLessonRepository deletedLessonRepository;
 
 
   @Override
@@ -39,9 +46,17 @@ public class AbonementServiceImpl implements AbonementService{
   }
 
   @Override
-  public void deleteAbonement(Long id) {
+  public ApiResponse deleteAbonement(Long id) {
     Abonement abonement = abonementRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Abonement", "id", id));
+    List<ConfirmedLesson> confirmedLessons = confirmedLessonRepository.findAllByAbonement(abonement);
+    List<DeletedLesson> deletedLessons = deletedLessonRepository.findAllByAbonement(abonement);
+
+    if (confirmedLessons.size() != 0 || deletedLessons.size() != 0) {
+      return new ApiResponse(false, "Не вдалося відалити, оскільки заняття пов'язані з цим абонементом");
+    }
+
     abonementRepository.delete(abonement);
+    return new ApiResponse(true, "Абонемент видалено");
   }
 
   @Override
